completed home page animation

+++ app/views/posts/index.blade.php
@@ -1,30 +1,45 @@
+<?php  
+	
+
+
+?>
+
 @extends('layouts.master')
+
+@section('insert_nav')	
+	<form method="get" action="{{{ action('PostsController@index') }}}">
+		<input id="search" type="search" class="navbar-text navbar-left" name="query" placeholder="search">
+	</form>
+@stop
+
 
 @section('content')
 	<h1>All the Blogs</h1>
 	@foreach ($allPosts as $post)
 		
-		<h3>{{{ $post['attributes']['title'] }}}</h3>
+		<a href="{{{ action('PostsController@show', ['id' => $post['attributes']['id']]) }}}"><h3>{{{ $post['attributes']['title'] }}}</h3></a>
+		<h3>{{{ $post->user->first_name . " " . $post->user->last_name }}}</h3>
 
-		{{ Form::open(['action' => ['PostsController@show', $post['attributes']['id']] , 'method'=>'GET']) }}
-			{{ Form::submit('View Post') }}
-		{{ Form::close() }}
-		{{ Form::open(['action' => ['PostsController@edit', $post['attributes']['id']] , 'method'=>'GET']) }}
-			{{ Form::submit('Edit Post') }}
-		{{ Form::close() }}
-		{{ Form::open(['action' => ['PostsController@destroy', $post['attributes']['id']] , 'method'=>'DELETE']) }}
-			{{ Form::submit('delete',['class'=>'btn btn-danger']) }}
-		{{ Form::close() }}
-
+		@if(Auth::check())
+			{{ Form::open(['action' => ['PostsController@edit', $post['attributes']['id']] , 'method'=>'GET']) }}
+				{{ Form::submit('Edit Post') }}
+			{{ Form::close() }}
+			{{ Form::open(['action' => ['PostsController@destroy', $post['attributes']['id']] , 'method'=>'DELETE']) }}
+				{{ Form::submit('delete',['class'=>'btn btn-danger']) }}
+			{{ Form::close() }}
+		@endif
 		 
 	@endforeach 
 
+	@if(Auth::check())
+		<form method="GET" action="{{{ action('PostsController@create') }}}">
+			<button type="submit">Create a New Post</button>
+		</form>
+	@endif
 
-	<form method="GET" action="{{{ action('PostsController@create') }}}">
-		<button type="submit">Create a New Post</button>
-	</form>
-
-	{{ $allPosts->links() }}
+	@if(false)
+		{{ $allPosts->links() }}
+	@endif
 
 @stop
 
@@ -36,6 +51,11 @@
 
 			var is_delete = confirm("are you sure you want to delte this post?");
 
+		});
+	    $('#search').keypress(function(e) {    
+            if(e.which == 10 || e.which == 13) {
+                this.form.submit();
+            }
 		});
 
 	</script>
